clean up the junk
