+++ cmd/transit/duboce/parser/structs.go
